+++ cocos/base/CCString.cpp
@@ -255,7 +255,7 @@ String* String::createWithFormat(const char* format, ...)
 
 String* String::createWithContentsOfFile(const char* filename)
 {
-    long size = 0;
+    size_t size = 0;
     unsigned char* data = 0;
     String* ret = NULL;
     data = FileUtils::getInstance()->getFileData(filename, "rb", &size);
